work on strings and memory functions

+++ include/gigatron/stdio.h
@@ -37,7 +37,7 @@ typedef long fpos_t;
 #define _SIZE_T
 #define _SIZE_T_
 #define _SIZE_T_DEFINED
-typedef unsigned long size_t;
+typedef unsigned int size_t;
 #endif
 
 #if !defined(_VA_LIST) && !defined(_VA_LIST_DEFINED)
